+++ packages/babel-preset-expo/src/use-dom-directive-plugin.ts
@@ -6,11 +6,12 @@ import crypto from 'crypto';
 import { basename } from 'path';
 import url from 'url';
 
-import { getIsProd } from './common';
+import { getIsDOM, getIsProd } from './common';
 
 export function expoUseDomDirectivePlugin(api: ConfigAPI): babel.PluginObj {
   // TODO: Is exporting
   const isProduction = api.caller(getIsProd);
+  const isDOMBundle = api.caller(getIsDOM);
   const platform = api.caller((caller) => (caller as any)?.platform);
 
   return {
@@ -18,7 +19,7 @@ export function expoUseDomDirectivePlugin(api: ConfigAPI): babel.PluginObj {
     visitor: {
       Program(path, state) {
         // Native only feature.
-        if (platform === 'web') {
+        if (platform === 'web' && !isDOMBundle) {
           return;
         }
 
@@ -36,6 +37,15 @@ export function expoUseDomDirectivePlugin(api: ConfigAPI): babel.PluginObj {
         // File starts with "use dom" directive.
         if (!hasUseDomDirective) {
           // Do nothing for code that isn't marked as a dom component.
+          return;
+        }
+
+        if (isDOMBundle) {
+          // Inject the DOM component registration.
+          path.pushContainer(
+            'body',
+            template.ast(`require('expo/dom/internal').registerDOMComponent(exports.default);`)
+          );
           return;
         }
 
@@ -67,6 +77,7 @@ export function expoUseDomDirectivePlugin(api: ConfigAPI): babel.PluginObj {
           `import React from 'react';
 import { WebView } from 'expo/dom/internal';`,
         ];
+
         if (isProduction) {
           // MUST MATCH THE EXPORT COMMAND!
           const hash = crypto.createHash('sha1').update(outputKey).digest('hex');